STORM-2845 Drop standalone mode of Storm SQL
* remove all related interfaces/classes on standalone mode
* migrate tests to trident mode which are associated to standalone mode
* remove comments from tests which Calcite has fixed the issues
