+++ sql/storm-sql-core/src/test/org/apache/storm/sql/compiler/backends/trident/TestPlanCompiler.java
@@ -22,18 +22,15 @@ package org.apache.storm.sql.compiler.backends.trident;
 import com.google.common.collect.ImmutableMap;
 import org.apache.calcite.DataContext;
 import org.apache.calcite.avatica.util.DateTimeUtils;
-import org.apache.storm.Config;
 import org.apache.storm.LocalCluster;
-import org.apache.storm.LocalCluster.LocalTopology;
+import org.apache.storm.sql.SqlTestUtil;
 import org.apache.storm.sql.TestUtils;
-import org.apache.storm.sql.javac.CompilingClassLoader;
 import org.apache.storm.sql.planner.trident.QueryPlanner;
 import org.apache.storm.sql.runtime.ISqlTridentDataSource;
 import org.apache.storm.sql.AbstractTridentProcessor;
 import org.apache.storm.trident.TridentTopology;
 import org.apache.storm.tuple.Fields;
 import org.apache.storm.tuple.Values;
-import org.apache.storm.utils.Utils;
 import org.junit.AfterClass;
 import org.junit.Assert;
 import org.junit.Before;
@@ -43,10 +40,11 @@ import org.junit.Test;
 import java.time.ZoneOffset;
 import java.util.Arrays;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
-import java.util.concurrent.Callable;
 
 import static org.apache.storm.sql.TestUtils.MockState.getCollectedValues;
+import static org.junit.Assert.assertEquals;
 
 public class TestPlanCompiler {
   private static LocalCluster cluster;
@@ -82,7 +80,7 @@ public class TestPlanCompiler {
     Fields f = proc.outputStream().getOutputFields();
     proc.outputStream().partitionPersist(new TestUtils.MockStateFactory(),
             f, new TestUtils.MockStateUpdater(), new Fields());
-    runTridentTopology(EXPECTED_VALUE_SIZE, proc, topo);
+    SqlTestUtil.runTridentTopology(cluster, EXPECTED_VALUE_SIZE, proc, topo);
     Assert.assertArrayEquals(new Values[] { new Values(3), new Values(4)}, getCollectedValues().toArray());
   }
 
@@ -98,7 +96,7 @@ public class TestPlanCompiler {
     QueryPlanner planner = new QueryPlanner(state.schema());
     AbstractTridentProcessor proc = planner.compile(data, sql);
     final TridentTopology topo = proc.build();
-    runTridentTopology(EXPECTED_VALUE_SIZE, proc, topo);
+    SqlTestUtil.runTridentTopology(cluster, EXPECTED_VALUE_SIZE, proc, topo);
     Assert.assertArrayEquals(new Values[] { new Values(4, "abcde", "y")}, getCollectedValues().toArray());
   }
 
@@ -118,28 +116,8 @@ public class TestPlanCompiler {
     Fields f = proc.outputStream().getOutputFields();
     proc.outputStream().partitionPersist(new TestUtils.MockStateFactory(),
             f, new TestUtils.MockStateUpdater(), new Fields());
-    runTridentTopology(EXPECTED_VALUE_SIZE, proc, topo);
+    SqlTestUtil.runTridentTopology(cluster, EXPECTED_VALUE_SIZE, proc, topo);
     Assert.assertArrayEquals(new Values[] { new Values(5) }, getCollectedValues().toArray());
-  }
-
-  @Test
-  public void testCaseStatement() throws Exception {
-    int EXPECTED_VALUE_SIZE = 5;
-    String sql = "SELECT CASE WHEN NAME IN ('a', 'abc', 'abcde') THEN UPPER('a') " +
-            "WHEN UPPER(NAME) = 'AB' THEN 'b' ELSE {fn CONCAT(NAME, '#')} END FROM FOO";
-    TestCompilerUtils.CalciteState state = TestCompilerUtils.sqlOverDummyTable(sql);
-
-    final Map<String, ISqlTridentDataSource> data = new HashMap<>();
-    data.put("FOO", new TestUtils.MockSqlTridentDataSource());
-
-    QueryPlanner planner = new QueryPlanner(state.schema());
-    AbstractTridentProcessor proc = planner.compile(data, sql);
-    final TridentTopology topo = proc.build();
-    Fields f = proc.outputStream().getOutputFields();
-    proc.outputStream().partitionPersist(new TestUtils.MockStateFactory(), f, new TestUtils.MockStateUpdater(), new Fields());
-    runTridentTopology(EXPECTED_VALUE_SIZE, proc, topo);
-
-    Assert.assertArrayEquals(new Values[]{new Values("A"), new Values("b"), new Values("A"), new Values("abcd#"), new Values("A")}, getCollectedValues().toArray());
   }
 
   @Test
@@ -158,18 +136,66 @@ public class TestPlanCompiler {
     final TridentTopology topo = proc.build();
     Fields f = proc.outputStream().getOutputFields();
     proc.outputStream().partitionPersist(new TestUtils.MockStateFactory(), f, new TestUtils.MockStateUpdater(), new Fields());
-    runTridentTopology(EXPECTED_VALUE_SIZE, proc, topo);
+    SqlTestUtil.runTridentTopology(cluster, EXPECTED_VALUE_SIZE, proc, topo);
 
     Map<String, Integer> map = ImmutableMap.of("b", 2, "c", 4);
     Map<String, Map<String, Integer>> nestedMap = ImmutableMap.of("a", map);
     Assert.assertArrayEquals(new Values[]{new Values(2, 4, nestedMap, Arrays.asList(100, 200, 300))}, getCollectedValues().toArray());
   }
 
+  /**
+   * All the binary literal tests are done here, because Avatica converts the result to byte[]
+   * whereas Trident provides the result to ByteString which makes different semantic from Trident implementation.
+   */
   @Test
-  public void testDateKeywords() throws Exception {
+  public void testBinaryStringFunctions() throws Exception {
+    int EXPECTED_VALUE_SIZE = 1;
+    String sql = "SELECT x'45F0AB' || x'45F0AB', " +
+            "POSITION(x'F0' IN x'453423F0ABBC'), " +
+            "OVERLAY(x'453423F0ABBC45' PLACING x'4534' FROM 3), " +
+            "SUBSTRING(x'453423F0ABBC' FROM 3), " +
+            "SUBSTRING(x'453423F0ABBC453423F0ABBC' FROM 3 FOR 4) " +
+            "FROM FOO " +
+            "WHERE ID > 0 AND ID < 2";
+
+    TestCompilerUtils.CalciteState state = TestCompilerUtils.sqlOverDummyTable(sql);
+
+    final Map<String, ISqlTridentDataSource> data = new HashMap<>();
+    data.put("FOO", new TestUtils.MockSqlTridentDataSource());
+    QueryPlanner planner = new QueryPlanner(state.schema());
+    AbstractTridentProcessor proc = planner.compile(data, sql);
+    final TridentTopology topo = proc.build();
+    Fields f = proc.outputStream().getOutputFields();
+    proc.outputStream().partitionPersist(new TestUtils.MockStateFactory(), f, new TestUtils.MockStateUpdater(), new Fields());
+    SqlTestUtil.runTridentTopology(cluster, EXPECTED_VALUE_SIZE, proc, topo);
+
+    List<Object> v = getCollectedValues().get(0);
+
+    assertEquals("45f0ab45f0ab", v.get(0).toString());
+    assertEquals(4, v.get(1));
+    assertEquals("45344534abbc45", v.get(2).toString());
+    assertEquals("23f0abbc", v.get(3).toString());
+    assertEquals("23f0abbc", v.get(4).toString());
+  }
+
+  /**
+   * All the date/time/timestamp related tests are done here, because Avatica converts the result of date functions to java.sql classes
+   * whereas Trident provides long type which makes different semantic from Trident implementation.
+   */
+  @Test
+  public void testDateKeywordsAndFunctions() throws Exception {
     int EXPECTED_VALUE_SIZE = 1;
     String sql = "SELECT " +
-            "LOCALTIME, CURRENT_TIME, LOCALTIMESTAMP, CURRENT_TIMESTAMP, CURRENT_DATE " +
+            "LOCALTIME, CURRENT_TIME, LOCALTIMESTAMP, CURRENT_TIMESTAMP, CURRENT_DATE, " +
+            "DATE '1970-05-15' AS datefield, TIME '00:00:00' AS timefield, TIMESTAMP '2016-01-01 00:00:00' as timestampfield, " +
+            "EXTRACT(MONTH FROM TIMESTAMP '2010-01-23 12:34:56')," +
+            "FLOOR(DATE '2016-01-23' TO MONTH)," +
+            "CEIL(TIME '12:34:56' TO MINUTE)," +
+            "{fn CURDATE()} = CURRENT_DATE, {fn CURTIME()} = LOCALTIME, {fn NOW()} = LOCALTIMESTAMP," +
+            "{fn QUARTER(DATE '2016-10-07')}, {fn TIMESTAMPADD(MINUTE, 15, TIMESTAMP '2016-10-07 00:00:00')}," +
+            "{fn TIMESTAMPDIFF(SECOND, TIMESTAMP '2016-10-06 00:00:00', TIMESTAMP '2016-10-07 00:00:00')}," +
+            "INTERVAL '1-5' YEAR TO MONTH AS intervalfield, " +
+            "(DATE '1970-01-01', DATE '1970-01-15') AS anchoredinterval_field "   +
             "FROM FOO " +
             "WHERE ID > 0 AND ID < 2";
     TestCompilerUtils.CalciteState state = TestCompilerUtils.sqlOverDummyTable(sql);
@@ -182,7 +208,7 @@ public class TestPlanCompiler {
     final TridentTopology topo = proc.build();
     Fields f = proc.outputStream().getOutputFields();
     proc.outputStream().partitionPersist(new TestUtils.MockStateFactory(), f, new TestUtils.MockStateUpdater(), new Fields());
-    runTridentTopology(EXPECTED_VALUE_SIZE, proc, topo);
+    SqlTestUtil.runTridentTopology(cluster, EXPECTED_VALUE_SIZE, proc, topo);
 
     long utcTimestamp = (long) dataContext.get(DataContext.Variable.UTC_TIMESTAMP.camelName);
     long currentTimestamp = (long) dataContext.get(DataContext.Variable.CURRENT_TIMESTAMP.camelName);
@@ -195,38 +221,9 @@ public class TestPlanCompiler {
     int localTimeInt = (int) (localTimestamp % DateTimeUtils.MILLIS_PER_DAY);
     int currentTimeInt = (int) (currentTimestamp % DateTimeUtils.MILLIS_PER_DAY);
 
-    Assert.assertArrayEquals(new Values[]{new Values(localTimeInt, currentTimeInt, localTimestamp, currentTimestamp, dateInt)}, getCollectedValues().toArray());
+    Assert.assertArrayEquals(new Values[]{new Values(localTimeInt, currentTimeInt, localTimestamp, currentTimestamp, dateInt,
+                    134, 0, 1451606400000L, 1L, 0L, 45300000, true, true, true, 4L, 1475799300000L, 86400, 17, 0, 14)},
+            getCollectedValues().toArray());
   }
 
-  private void runTridentTopology(final int expectedValueSize, AbstractTridentProcessor proc,
-                                  TridentTopology topo) throws Exception {
-    final Config conf = new Config();
-    conf.setMaxSpoutPending(20);
-
-    if (proc.getClassLoaders() != null && proc.getClassLoaders().size() > 0) {
-      CompilingClassLoader lastClassloader = proc.getClassLoaders().get(proc.getClassLoaders().size() - 1);
-      Utils.setClassLoaderForJavaDeSerialize(lastClassloader);
-    }
-
-    try (LocalTopology stormTopo = cluster.submitTopology("storm-sql", conf, topo.build())) {
-      waitForCompletion(1000 * 1000, new Callable<Boolean>() {
-        @Override
-        public Boolean call() throws Exception {
-          return getCollectedValues().size() < expectedValueSize;
-        }
-      });
-    } finally {
-      while(cluster.getClusterInfo().get_topologies_size() > 0) {
-        Thread.sleep(10);
-      }
-      Utils.resetClassLoaderForJavaDeSerialize();
-    }
-  }
-
-  private void waitForCompletion(long timeout, Callable<Boolean> cond) throws Exception {
-    long start = TestUtils.monotonicNow();
-    while (TestUtils.monotonicNow() - start < timeout && cond.call()) {
-      Thread.sleep(100);
-    }
-  }
 }
