+++ sql/storm-sql-runtime/src/jvm/org/apache/storm/sql/runtime/datasource/socket/SocketDataSourcesProvider.java
@@ -23,7 +23,6 @@ import java.util.List;
 import java.util.Properties;
 
 import org.apache.storm.spout.Scheme;
-import org.apache.storm.sql.runtime.DataSource;
 import org.apache.storm.sql.runtime.DataSourcesProvider;
 import org.apache.storm.sql.runtime.FieldInfo;
 import org.apache.storm.sql.runtime.IOutputSerializer;
@@ -76,11 +75,6 @@ public class SocketDataSourcesProvider implements DataSourcesProvider {
             StateUpdater<SocketState> stateUpdater = new SocketStateUpdater(serializer);
             return new SimpleSqlTridentConsumer(stateFactory, stateUpdater);
         }
-    }
-
-    @Override
-    public DataSource construct(URI uri, String inputFormatClass, String outputFormatClass, List<FieldInfo> fields) {
-        throw new UnsupportedOperationException();
     }
 
     @Override
